requirements.txt
Django==5.1.7
djangorestframework==3.15.0
psycopg2==2.9.6  # PostgreSQL adapter for Django
python-dotenv==1.0.0  # For loading environment variables securely
drf-yasg==1.21.3  # For Swagger API documentation
